PP-634: Fully refactored settings
